requirements.txt
altcha==0.1.3
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.13.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
libsass==0.23.0
MarkupSafe==2.1.5
mongoengine==0.28.2
msgspec==0.18.6
openai==1.25.1
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.6.3
python-dotenv==1.0.1
redis==5.0.4
regex==2024.4.28
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.30
stripe==9.11.0
tiktoken==0.9.0
tqdm==4.66.4
typing_extensions==4.11.0
ulid-py==1.1.0
urllib3==2.2.1
valkey==5.1.0b7
Werkzeug==3.0.2
